[INJIMOB-1303] rename vci client package
