+++ android/app/src/main/java/io/mosip/residentapp/InjiPackage.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class MyInjiPackage implements ReactPackage {
+public class InjiPackage implements ReactPackage {
     @NonNull
     @Override
     public List<NativeModule> createNativeModules(@NonNull ReactApplicationContext reactApplicationContext) {
