+++ machines/Issuers/IssuersService.ts
@@ -4,8 +4,8 @@ import NetInfo from '@react-native-community/netinfo';
 import {request} from '../../shared/request';
 import {
   constructAuthorizationConfiguration,
+  constructProofJWT,
   getCredentialType,
-  getJWK,
   Issuers,
   Issuers_Key_Ref,
   updateCredentialInformation,
@@ -14,7 +14,6 @@ import {
 import {authorize} from 'react-native-app-auth';
 import {
   generateKeys,
-  getJWT,
   isHardwareKeystoreExists,
 } from '../../shared/cryptoutil/cryptoUtil';
 import SecureKeystore from '@mosip/secure-keystore';
@@ -29,7 +28,6 @@ import {
 } from '../../shared/telemetry/TelemetryUtils';
 import {TelemetryConstants} from '../../shared/telemetry/TelemetryConstants';
 import {VciClient} from '../../shared/vciClient/VciClient';
-import jwtDecode from 'jwt-decode';
 
 export const IssuersService = () => {
   return {
@@ -69,25 +67,11 @@ export const IssuersService = () => {
         credentialType: getCredentialType(context),
         credentialFormat: 'ldp_vc',
       };
-      const jwtHeader = {
-        alg: 'RS256',
-        jwk: await getJWK(context.publicKey),
-        typ: 'openid4vci-proof+jwt',
-      };
-      const decodedToken = jwtDecode(context.tokenResponse.accessToken);
-      const jwtPayload = {
-        iss: context.selectedIssuer.client_id,
-        nonce: decodedToken.c_nonce,
-        aud: context.selectedIssuer.credential_audience,
-        iat: Math.floor(new Date().getTime() / 1000),
-        exp: Math.floor(new Date().getTime() / 1000) + 18000,
-      };
-
-      const proofJWT = await getJWT(
-        jwtHeader,
-        jwtPayload,
-        Issuers_Key_Ref,
+      const proofJWT = await constructProofJWT(
+        context.publicKey,
         context.privateKey,
+        accessToken,
+        context.selectedIssuer,
       );
       let credential = await VciClient.downloadCredential(
         issuerMeta,
